implement the influxdb cache mechanism

+++ src/common/slurm_protocol_defs.h
@@ -736,17 +736,19 @@ typedef struct step_complete_msg {
  	uint32_t step_rc;	/* largest task return code */
 	jobacctinfo_t *jobacct;
 } step_complete_msg_t;
+
 #ifdef __METASTACK_LOAD_ABNORMAL
 typedef struct step_gather_msg {
 	uint32_t rank;	/* First node rank within job step's alloc */
 	//uint32_t depth_child; /*The current structure contains all child nodes*/
 	slurm_step_id_t step_id; /* Job step ID */
-	uint64_t cpu_util;         /* Real-time utilization of individual job steps on the node */
-	uint64_t cpu_ave;       	 /* Average utilization of individual job steps on the node */
+	double cpu_util;         /* Real-time utilization of individual job steps on the node */
+	double cpu_ave;       	 /* Average utilization of individual job steps on the node */
 	uint64_t mem_real;  		 /* Real-time memory utilization of individual job steps on the node*/
 	uint64_t vmem_real;          /* Average memory utilization of individual job steps on the node */
 	uint64_t page_fault;     /* The total number of page fault exceptions in individual job steps on the node */
 	uint64_t load_flag; 
+	
 	//bool pid_status;
 } step_gather_msg_t;
 #endif
